# docs/openapi/pipeline.yaml
openapi: 3.0.0
info:
  description: Pipeline v0.3.0 swagger
  version: "0.3.0"
  title: Pipeline API
  contact:
    email: info@banzaicloud.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: http://localhost:9090
...
paths:
...
  '/api/v1/orgs/{orgId}/helm/repos':
    get:
      security:
        - bearerAuth: []
      tags:
       - helm
      summary: List repositories
      operationId: HelmInit
      description: Listing Helm repositories in the cluster
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization identification
          schema:
            type: integer
      responses:
        '200':
          description: "Repositories listed"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelmReposListResponse'
        '400':
          description: "Error during getting endpoints"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError_400'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: "Cluster not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterNotFound'

    post:
      security:
        - bearerAuth: []
      tags:
       - helm
      summary: Add Repo
      operationId: HelmReposAdd
      description: Add new Helm repository
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization identification
          schema:
            type: integer
      responses:
        '200':
          description: "resource successfully added."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelmRepoListItem'
        '400':
          description: "Error adding Helm repo"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError_400'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: "Cluster not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterNotFound'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelmReposAddRequest'
  '/api/v1/orgs/{orgId}/helm/repos/{repoName}':
    delete:
      security:
        - bearerAuth: []
      tags:
       - helm
      summary: Delete Repo
      operationId: HelmReposDelete
      description: Delete  Helm repository
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization identification
          schema:
            type: integer
        - name: repoName
          in: path
          required: true
          description: Helm repo name
          schema:
            type: string
      responses:
        '200':
          description: "resource deleted successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelmReposDeleteResponse'
        '400':
          description: "error deleting Helm repos"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError_400'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: "Cluster not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterNotFound'

    put:
      security:
        - bearerAuth: []
      tags:
       - helm
      summary: Modify Repo
      operationId: HelmReposModify
      description: Modify Helm repository
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization identification
          schema:
            type: integer
        - name: repoName
          in: path
          required: true
          description: Helm repo name
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelmReposModifyRequest'
      responses:
        '200':
          description: "resource modified successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelmReposUpdateResponse'
        '400':
          description: "repo modification failed"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError_400'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: "Resource found"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ClusterNotFound'
                  - $ref: '#/components/schemas/RepoNotFound'
  '/api/v1/orgs/{orgId}/helm/repos/{repoName}/update':
    put:
      security:
        - bearerAuth: []
      tags:
       - helm
      summary: Update Repo
      operationId: HelmReposUpdate
      description: Modify Helm repository
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization identification
          schema:
            type: integer
        - name: repoName
          in: path
          required: true
          description: Helm repo name
          schema:
            type: string
      responses:
        '200':
          description: "resource modified successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelmReposUpdateResponse'
        '400':
          description: "repo modification failed"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError_400'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: "Resource found"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ClusterNotFound'
                  - $ref: '#/components/schemas/RepoNotFound'
...
  '/api/v1/orgs/{orgId}/clusters/{id}/deployments':
    get:
      security:
        - bearerAuth: []
      tags:
        - deployments
      summary: List deployments
      operationId: ListDeployments
      description: Listing Helm deployments
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization identification
          schema:
            type: integer
        - name: id
          in: path
          required: true
          description: Selected cluster identification (number)
          schema:
            type: integer
      responses:
        '200':
          description: "Deployments listed"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListDeploymentsResponse'
        '400':
          description: "Error during getting endpoints"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError_400'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: "Cluster not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterNotFound'
    post:
      security:
        - bearerAuth: []
      tags:
        - deployments
      summary: Create a Helm deployment
      operationId: CreateDeployment
      description: Creating a Helm deployment
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization identification
          schema:
            type: integer
        - name: id
          in: path
          required: true
          description: Selected cluster identification (number)
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUpdateDeploymentRequest'
      responses:
        '201':
          description: "Deployment created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUpdateDeploymentResponse'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '400':
          description: "Error during creating deployment"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseError_400'
        '404':
          description: "Cluster not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterNotFound'
    head:
      security:
        - bearerAuth: []
      tags:
        - deployments
      summary: Get tiller status
      operationId: GetTillerStatus
      description: Checking if tiller ready to accept deployments
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization identification
          schema:
            type: integer
        - name: id
          in: path
          required: true
          description: Selected cluster identification (number)
          schema:
            type: integer
      responses:
        '200':
          description: "Tiller is available"
        '400':
          description: "Error during creating deployment"
        '401':
          description: "Unauthorized"
        '404':
          description: "Cluster not found"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaseError_400:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "Error during process"
        error:
          type: string
          example: "Error during process"
...
    ClusterNotFound:
      type: object
      properties:
        code:
          type: integer
          example: 404
        message:
          type: string
          example: "Cluster not found"
        error:
          type: string
          example: "record not found"
    RepoNotFound:
      type: object
      properties:
        code:
          type: integer
          example: 404
        message:
          type: string
          example: "repo not found"
        error:
          type: string
          example: "repo not found"
...
    Unauthorized:
      type: object
      properties:
        code:
          type: integer
          example: 401
        message:
          type: string
          example: "Invalid token"
        error:
          type: string
          example: "token contains an invalid number of segments"
...
    ListDeploymentsResponse:
      type: array
      items:
        type: object
        properties:
          releaseName:
            type: string
            example: "pipeline"
          chart:
            type: string
            example: "pipeline-cluster-ingress-0.0.1"
          chartName:
            type: string
            example: "pipeline-cluster-ingress"
          chartVersion:
            type: string
            example: "0.0.1"
          version:
            type: integer
            example: 1
          updatedAt:
            type: string
            example: "2018-07-03T14:23:19+02:00"
          status:
            type: string
            example: "DEPLOYED"
          namespace:
            type: string
            example: "default"
          createdAt:
            type: string
            example: "2018-07-03T14:23:19+02:00"
    CreateUpdateDeploymentRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "banzaicloud-stable/drone"
        version:
          type: string
          example: "0.1.0"
          description: "Version of the deployment. If not specified, the latest version is used."
        namespace:
          type: string
          example: "default"
        releaseName:
          type: string
          example: "singed-bee"
        reuseValues:
          type: boolean
          example: "true"
        values:
          type: object
          additionalProperties: true
          description: current values of the deployment
          example: { "ingress": { "enabled": "true" } }
...
    HelmReposListResponse:
      type: array
      items:
        $ref: '#/components/schemas/HelmRepoListItem'
    HelmRepoListItem:
      type: object
      properties:
        name:
          type: string
          example: "stable"
        cache:
          type: string
          example: "statestore/<cluster_name>/helm/repository/cache/stabl-index.yaml"
        url:
          type: string
          example: "https://kubernetes-charts.storage.googleapis.com"
        certFile:
          type: string
          example: ""
        keyFile:
          type: string
          example: ""
        caFile:
          type: string
          example: ""
...
    HelmReposModifyRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        certFile:
          type: string
        keyFile:
          type: string
        caFile:
          type: string
      example:
          url: "https://kubernetes-charts.storage.googleapis.com"
    HelmReposAddRequest:
      type: object
      required:
        - name
        - url
      properties:
        name:
          type: string
        url:
          type: string
        certFile:
          type: string
        keyFile:
          type: string
        caFile:
          type: string
      example:
          name: "stable"
          url: "https://kubernetes-charts.storage.googleapis.com"
    HelmReposDeleteResponse:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: "resource deleted successfully"
        name:
          type: string
          example: "stable"
    HelmReposUpdateResponse:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: "resource modified successfully"
        name:
          type: string
          example: "stable"
...